feat: Add ValidateToken (#1)

+++ src/main/java/com/swugether/server/db/dao/RefreshTokenRepository.java
@@ -3,5 +3,5 @@ package com.swugether.server.db.dao;
 import com.swugether.server.db.domain.RefreshTokenEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RefreshTokenRepository extends CrudRepository<RefreshTokenEntity, Integer> {
+public interface RefreshTokenRepository extends CrudRepository<RefreshTokenEntity, Long> {
 }
